Move trait definition and add comments
